move validation funcs into their own file
